DEV-5523: use mixin to express 409 conflict when resource is unique
