+++ apps/contributions/tests/views/test_switchboard.py
@@ -583,7 +583,7 @@ class TestSwitchboardPaymentsViewSet:
         response = api_client.post(
             reverse("switchboard-payment-list"), data=payment_creation_data, headers={"Authorization": f"Token {token}"}
         )
-        assert response.status_code == status.HTTP_400_BAD_REQUEST
+        assert response.status_code == status.HTTP_409_CONFLICT
 
     def test_retrieve(self, api_client, payment, token):
         response = api_client.get(
@@ -647,7 +647,7 @@ class TestSwitchboardPaymentsViewSet:
             headers={"Authorization": f"Token {token}"},
             format="json",
         )
-        assert response.status_code == status.HTTP_400_BAD_REQUEST
+        assert response.status_code == status.HTTP_409_CONFLICT
         assert response.json() == {
             "stripe_balance_transaction_id": ["payment with this stripe balance transaction id already exists."]
         }
